+++ src/libraries/qlitehtml/src/3rdparty/litehtml/src/gumbo/include/gumbo/utf8.h
@@ -38,91 +38,94 @@
 #include "gumbo.h"
 
 #ifdef __cplusplus
-extern "C" {
+extern "C"
+{
 #endif
 
-struct GumboInternalError;
-struct GumboInternalParser;
+	struct GumboInternalError;
+	struct GumboInternalParser;
 
-// Unicode replacement char.
-extern const int kUtf8ReplacementChar;
+	// Unicode replacement char.
+	extern const int kUtf8ReplacementChar;
 
-typedef struct GumboInternalUtf8Iterator {
-  // Points at the start of the code point most recently read into 'current'.
-  const char* _start;
+	typedef struct GumboInternalUtf8Iterator
+	{
+		// Points at the start of the code point most recently read into 'current'.
+		const char* _start;
 
-  // Points at the mark.  The mark is initially set to the beginning of the
-  // input.
-  const char* _mark;
+		// Points at the mark.  The mark is initially set to the beginning of the
+		// input.
+		const char* _mark;
 
-  // Points past the end of the iter, like a past-the-end iterator in the STL.
-  const char* _end;
+		// Points past the end of the iter, like a past-the-end iterator in the STL.
+		const char* _end;
 
-  // The code point under the cursor.
-  int _current;
+		// The code point under the cursor.
+		int _current;
 
-  // The width in bytes of the current code point.
-  ptrdiff_t _width;
+		// The width in bytes of the current code point.
+		int _width;
 
-  // The SourcePosition for the current location.
-  GumboSourcePosition _pos;
+		// The SourcePosition for the current location.
+		GumboSourcePosition _pos;
 
-  // The SourcePosition for the mark.
-  GumboSourcePosition _mark_pos;
+		// The SourcePosition for the mark.
+		GumboSourcePosition _mark_pos;
 
-  // Pointer back to the GumboParser instance, for configuration options and
-  // error recording.
-  struct GumboInternalParser* _parser;
-} Utf8Iterator;
+		// Pointer back to the GumboParser instance, for configuration options and
+		// error recording.
+		struct GumboInternalParser* _parser;
+	} Utf8Iterator;
 
-// Returns true if this Unicode code point is in the list of characters
-// forbidden by the HTML5 spec, such as NUL bytes and undefined control chars.
-bool utf8_is_invalid_code_point(int c);
+	// Returns true if this Unicode code point is in the list of characters
+	// forbidden by the HTML5 spec, such as NUL bytes and undefined control chars.
+	bool utf8_is_invalid_code_point(int c);
 
-// Initializes a new Utf8Iterator from the given byte buffer.  The source does
-// not have to be NUL-terminated, but the length must be passed in explicitly.
-void utf8iterator_init(struct GumboInternalParser* parser, const char* source, size_t source_length, Utf8Iterator* iter);
+	// Initializes a new Utf8Iterator from the given byte buffer.  The source does
+	// not have to be NUL-terminated, but the length must be passed in explicitly.
+	void utf8iterator_init(struct GumboInternalParser* parser, const char* source, size_t source_length,
+						   Utf8Iterator* iter);
 
-// Advances the current position by one code point.
-void utf8iterator_next(Utf8Iterator* iter);
+	// Advances the current position by one code point.
+	void utf8iterator_next(Utf8Iterator* iter);
 
-// Returns the current code point as an integer.
-int utf8iterator_current(const Utf8Iterator* iter);
+	// Returns the current code point as an integer.
+	int utf8iterator_current(const Utf8Iterator* iter);
 
-// Retrieves and fills the output parameter with the current source position.
-void utf8iterator_get_position(const Utf8Iterator* iter, GumboSourcePosition* output);
+	// Retrieves and fills the output parameter with the current source position.
+	void utf8iterator_get_position(const Utf8Iterator* iter, GumboSourcePosition* output);
 
-// Retrieves a character pointer to the start of the current character.
-const char* utf8iterator_get_char_pointer(const Utf8Iterator* iter);
+	// Retrieves a character pointer to the start of the current character.
+	const char* utf8iterator_get_char_pointer(const Utf8Iterator* iter);
 
-// Retrieves a character pointer to 1 past the end of the buffer.  This is
-// necessary for certain state machines and string comparisons that would like
-// to look directly for ASCII text in the buffer without going through the
-// decoder.
-const char* utf8iterator_get_end_pointer(const Utf8Iterator* iter);
+	// Retrieves a character pointer to 1 past the end of the buffer.  This is
+	// necessary for certain state machines and string comparisons that would like
+	// to look directly for ASCII text in the buffer without going through the
+	// decoder.
+	const char* utf8iterator_get_end_pointer(const Utf8Iterator* iter);
 
-// If the upcoming text in the buffer matches the specified prefix (which has
-// length 'length'), consume it and return true.  Otherwise, return false with
-// no other effects.  If the length of the string would overflow the buffer,
-// this returns false.  Note that prefix should not contain null bytes because
-// of the use of strncmp/strncasecmp internally.  All existing use-cases adhere
-// to this.
-bool utf8iterator_maybe_consume_match(Utf8Iterator* iter, const char* prefix, size_t length, bool case_sensitive);
+	// If the upcoming text in the buffer matches the specified prefix (which has
+	// length 'length'), consume it and return true.  Otherwise, return false with
+	// no other effects.  If the length of the string would overflow the buffer,
+	// this returns false.  Note that prefix should not contain null bytes because
+	// of the use of strncmp/strncasecmp internally.  All existing use-cases adhere
+	// to this.
+	bool utf8iterator_maybe_consume_match(Utf8Iterator* iter, const char* prefix, size_t length, bool case_sensitive);
 
-// "Marks" a particular location of interest in the input stream, so that it can
-// later be reset() to.  There's also the ability to record an error at the
-// point that was marked, as oftentimes that's more useful than the last
-// character before the error was detected.
-void utf8iterator_mark(Utf8Iterator* iter);
+	// "Marks" a particular location of interest in the input stream, so that it can
+	// later be reset() to.  There's also the ability to record an error at the
+	// point that was marked, as oftentimes that's more useful than the last
+	// character before the error was detected.
+	void utf8iterator_mark(Utf8Iterator* iter);
 
-// Returns the current input stream position to the mark.
-void utf8iterator_reset(Utf8Iterator* iter);
+	// Returns the current input stream position to the mark.
+	void utf8iterator_reset(Utf8Iterator* iter);
 
-// Sets the position and original text fields of an error to the value at the
-// mark.
-void utf8iterator_fill_error_at_mark(Utf8Iterator* iter, struct GumboInternalError* error);
+	// Sets the position and original text fields of an error to the value at the
+	// mark.
+	void utf8iterator_fill_error_at_mark(Utf8Iterator* iter, struct GumboInternalError* error);
 
 #ifdef __cplusplus
 }
 #endif
-#endif  // GUMBO_UTF8_H_
+#endif // GUMBO_UTF8_H_
